Database: make instances of OutcomeBuilder non copyable
The data is stored into the database by the dtor. If multiple copies
of the same instance exists, then the same data would be stored
many times, once for each copy. Moreover, as each instace may be
modified independently from the others, the schema used to store the
data may be different in each instance, ultimately causing a logic
error in the SQLite3 driver.